reactor: Add sugar to live-update options on all shards
The program_options updateable_value doesn't know if the option should
be smp::invoke_on_all-updated or not, it just calls the function.
Reactor options are mostly per-shard and deserve a common helper to
spread the option updating across shards. Here's the helper that will be
used by next patches

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>