resource: take kernel min_free_kbytes into account when allocating memory
The kernel declares OOM when memory dips below /proc/sys/vm/min_free_kbytes.
On normal systems, this is a reasonable value (69MB on my 128GB desktop)
and is far below the 7% we already reserve for the kernel. But on ARM
systems with 64kB pages and transparent hugepages enabled, this becomes
very large (~800MB on a 16GB system), presumably so there is space to
collect a 512MB hugepage. This reservation is already 5% of total system
memory, and accounting for other things, the system will OOM as soon
as Seastar allocates all of its memory. With prefault, this is within
a few seconds of starting, as seen in [1].

Fix this by subtracting min_free_kbytes from available memory. On
sane systems this has negligible effect, and on systems with very
large hugepage sizes, this will allow them to avoid running out of
memory.

[1] https://github.com/scylladb/scylladb/issues/14721

Closes #1747