build: do not use serial number file when signing a cert
after 18f43a0c7b, we always generate the certs created by
`mtls_certs` target for testing even if they already exist. this
is annoying. and can be misleading sometimes. as in general,
a build system does not need to rebuild its targets if the
targets are considered up-to-date.

in 18f43a0c7b, we add mtls_ca.srl as a DEPENDS of the
`add_custom_command()` calls which signs the x.509 certs.
quoting the part of the commit message:

> Also add the implicit dependency into the command declarations
> this doesn't help much, since the dependency is an input/output
> file which can't be expressed by the build system, but helps
> make the file visible.

and quoting the relative part in OPENSSL-X509(1):

> -CAcreateserial
>     With this option and the -CA option the CA serial number file
>     is created if it does not exist.  A random number is generated,
>     used for the certificate, and saved into the serial number file
>     determined as described above.

so strictly speaking the `openssl x509` command does not depend on
this serial number file in the sense that it can complete without
it.

in this this change,

* these new DEPENDS are removed from the the `add_custom_command()`
  calls.
* create a helper function for creating/signing certificates, so
  that is more obvious what the test intents to do.
* instead of using a CA serial number file, specify the serial number
  manually. since we only have 3 cert to sign, the serial number is
  hardwired instead of using `MATH()` command. it's overkill in this
  use case.
* "openssl_serial" job pool is dropped, as we don't to track the
  serial number with a file, and there is no need to serialize the
  write access to this file anymore.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>