io_queue, tests: Make tickets externally
This is a cosmetic change towards queued_io_request from io_queue.cc and
relevant queued-entities in tests. The change is in providing fair queue
ticket by the caller into the "entity" constructor, rather than latting
the latter construct it on its own. This change will be needed later,
when the ticket will be replaced with capacity and the capacity itself
will need data from the caller to be evaluated

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>