fair_queue: Introduce fair_queue_entry::capacity_t
This type is the "measure" of a queued entry. Right now it's the
fair_queue_ticket and all callers silently assume that. Later the
"measure" will become the fair_queue::capacity_t. This patch prepares
the rest of the code to use the fair_queue_entry::capacity_t instead of
fair_queue_ticket directly mainly to reduce the future churn

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>