fair_queue: Keep capacity_t on fair_queue_entry
Before this patch the "measure" of the entry was fair queue ticket. The
rest of the fair queue code always had to convert ticket to capacity,
because it's the capacity that's really used in fair-queue internals for
both -- fairness balancing and dispatch throttling.

This patch replaces the ticket with the capacity on entry thus making
on-demand conversions not needed and requiting callers to do this
conversion on their own beforehand

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>