io_queue, fair_queue: Calculate request tokens instead of ticket
The formula for request capacity is

  cap = weight / weight_max + size / size_max

where ..._max are constants currently sitting in fair queue.

This formula is implemented as

  cap = ticket(weight, size).normalize(ticket(weight_max / rate, size_max / rate)) * factor

where "rate" is the token bucket configuration rate set to std::milli by
the fair-queue code and factor is thd 1<<24 value to convert floating
point tokens into integer capacities.

This patch makes the formula more obvious in the implementation by
making the io-queue code calculate the floating point tokens according
to the former formula and the calling fair_queue::tokens_capacity() to
perform the needed rate and factor normalization.

This change slightly changes the way rounding works, so the actual
values of requests' capacities change. The change is +/-2 for the ~10k
capacity value, so not critical

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>