tests: Improve websocket parser tests
Currently websocket parser tests are done within general websocket
server test that involves network operations. This is wrong way to test
websocket parser because it involves unnecessary code and is harder to
control. For example, currently there is a race condition which is
resolved by sleeping by 100ms.

The solution is to extract the parser test to separate function. The
test is additionally improved to test more than one websocket payload at
the same time.

Co-authored-by: Dave Voutila <voutilad@gmail.com>

Closes scylladb/seastar#2562