io_queue: Add flow-rate based self slowdown backlink
Make the slow-down link less immediate. When the disk is not overloaded
it, in theory, should complete as many requests per-tick, as it was
dispatched to it previously. IOW, the ratio of dispatched requests to
completed requests should be 1.0 or, in the worst case, close to it.
In the "long run" it should be 1.0. If it's not, it means that requests
tend to accumulate in disk and it's a reason for slow down.

This ratio is tracked with the flow-rate monitor. It estimates the
ratio with the exponential moving average filter, adding new drops
once every 100ms with the smothing factor of 0.95.

Next, current formula for the rate limiter is

    IOPS / IOPS_max + Bandwidth / Bandwidth_max <= 1.0

and it's implemented as a token bucket with replenish rate of 1 token
per second and each request requiring

    capacity = weight / A + size / B

tokens (fraction) to proceed.  To slow things down there are two options:

 - reduce the 1.0 constant
 - increase capacities of requests

This set walks the second route, as touching the t.b. rate might be too
tricky, because t.b. is shared across shards and shards would step on
each others toes doing that. So instead of doing it, the patch increases
the requests' cost multiplying it by the aforementioned ratio. Since in
real load the ratio is always strictly greater than 1.0, a threshold is
added below which the scheduler continues working at the configured
rate.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>