io_queue: Add flow-rate based self slowdown backlink
As described in #1766, current attempt to make IO queue slowdown if disk
cannot keep up with the load is problematic. The main problem is that
the backlink relies on requests completions arriving in a timely manner
thus replenishing tokens into the bucket. It doesn't work well, because
the case when "disk is overloaded and delays completion of requests" is
indistinguishable from "reactor is overloaded and cannot poll
completions in a timely manner". As a result, lowering the rate of
dispatch on softly stalling reactor doesn't necessarily help making the
disk more happy. Reactor would continue delaying completions and the
backling would drop the rate even further.

The proposal is to make the slow-down link less immediate. When the disk
is not overloaded it, in theory, should complete as many requests
per-tick, as it was dispatched to it previously. IOW, the ratio of
dispatched requests to completed requests should be 1.0 or, in the worst
case, close to it. In the "long run" it should be 1.0. If it's not, it
means that requests tend to accumulate in disk and it's a reason for
slow down.

This ratio is tracked in #1766 PR with the flow-rate monitor. It
estimates the ratio with the exponential moving average filter, adding
new drops once every 100ms with the smothing factor of 0.95.

Next, current formula for the rate limiter is

    IOPS / IOPS_max + Bandwidth / Bandwidth_max <= 1.0

and it's implemented as a token bucket with replenish rate of 1 token
per second and each request requiring

    capacity = weight / A + size / B

tokens (fraction) to proceed.  To slow things down there are two options:

 - reduce the 1.0 constant
 - increase capacities of requests

This set walks the second route, as touching the t.b. rate might be too
tricky, because t.b. is shared across shards and shards would step on
each others toes doing that. So instead of doing it, the patch increases
the requests' cost multiplying it by the aforementioned ratio. Since in
real load the ratio is always strictly greater than 1.0, a threshold is
added below which the scheduler continues working at the configured
rate.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>