tests: Extern RPC stream tests to shutdown server "on the fly"
There's a bunch of test cases that validate how RPC stream is stopped in
case of normal closure and various injected errors. Add one more to test
how things stop when the server is shutdown() in the middle.

Note: this test would also pass without the previous changes, because,
as was described, closing server-side stream-parent would eventually
propagate to the stream itself. However, with the previous patches the
closure code still works, but happens some ticks earlier

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>