rpc: Add seastar::rpc::metrics
The class is in charge of maintaining rpc client metrics. There are two
extreme ways on how to export them.

First is to make each connection export metrics about itself. It's
possible, but since typically there are A LOT of connections, this would
explode the reported metrics list. Also, as connections appear and die
dynamically, it will be inconvenient to collect them on the client-side.

Second option is to sum up all the connections' counters into one place.
That's not great either, e.g. in multi-datacenter set-ups latency
metrics for intra-DC communications would be much better than those for
cross-DC ones. Mixing them together is producing some useless "average".
Another reason is that scylla uses different connections for RPC of
different "importance" to avoid head-of-line blocking problem. It's much
more informative to get separate sent/recvd/errors stats.

Having said that, the proposal is to introduce an aggregation of
connections based on the external configuration. RPC user would then
tell which "domain" a client belongs to and this will make metrics
aggregate per-connection metrics into a per-domain ones.

So first thing the new class does is collects the clients linked in
per-domain lists.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>