epoll: Avoid spinning on aborted connections
Using the epoll backend and having a non-closed socket that got terminated
with either EPOLLHUP|EPOLLERR under the hood but isn't closed yet can lead to
busy spinning in the reactor at 100%.

It's caused by the following scenario:

 - epoll_pwait returns an event for a socket with EPOLLIN|EPOLLHUP (or
   EPOLLOUT in a different respective case)
 - We enter the error branch and set evt.events. This doesn't make much
   difference at this point as pfd->events_requested is still equal to
   EPOLLIN
 - events_to_remove is calculated to 0
 - complete_epoll_event is called setting pfd->events_requested to 0
 - The loop concludes without removing the socket from the epoll set as
   we calculate events_to_remove before changing it. I think this is
   intended to avoid unneeded syscalls in the good case (so usually only
   receiving an event when not waiting for it will remove the socket from
   the set)
 - Given we run in level triggered mode epoll_wait again returns with
   EPOLLIN|EPOLLHUP for our socket
 - This time around in the error branch we set evt.events to 0
 - This causes events_to_remove again to be calculated to 0 even though
   pfd->events_requested is 0 as well
 - Above continues until the socket is eventually closed and hence
   implicitly removed from the epoll set

This patch prevents this from happening by pro-actively removing the fd
from the epoll set once we have received an error and are no longer
waiting for any events on the socket.

Note this should be relatively safe in any case. In case we ever get
interested in events again we would just add it back.

Fixes https://github.com/scylladb/seastar/issues/1785