prometheus.cc add native histogram
Prometheus native histograms are an alternative dense representation of a histogram.
They hold the benefits of high resolution at a lower resource cost.
This patch adds an option to send histograms in a native histogram
format.

To send a histogram in a native histogram format, all the following
conditions must be met:

* The allow_protobuf configuration should be set to true
* The requesting server should ask for a protobuf representation
* The histogram should include schema parameter