test: remove const from seastar_test::run_test_case()
remove the `const` specifier from `seastar_test::run_test_case()`,

`run_test_case()` is used by `seastar_test::run()` for running test.
before this change, we don't expect the test to mutate the
`seastar_test` instance instantiated for the test. but if we want to
use it as a fixture of test, `run_test_case()` would likely to mutate
the state of the instance of the class derived from `seastar_test`.
this would require us either to use a different way to drive the test,
or to change the function signature of `seastar_test::run_test_case()`.

Boost.Test uses a dedicated invoker for driving the test, while Seastar
use `seastar_test::run()`. to avoid the massive change, let's go
with a mutable `run_test_case()`. so, in this change, all the
`const` specifiers in function signatures of `run_test_case()` are
removed. this paves the road to a seasetar::testing with fixture support.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>