test: add fixture support to seastar::testing
before this change, because of the missing support of test fixture,
some tests in scylladb are using lambdas or homebrew test helper
class for hosting the helpers and for creating the test fixture,
so they can be shared in a test case or across test cases. this
hurts the readability. the better practice would be to have smaller
tests which share the same fixture explicitly.

so, in this change, two more macros are added so tests can use
fixture:

- SEASTAR_FIXTURE_TEST_CASE
- SEASTAR_THREAD_FIXTURE_TEST_CASE

tests/unit/tls_test.cc is updated to demostrate the use of
SEASTAR_THREAD_FIXTURE_TEST_CASE.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>