chunked_fifo: let incremetal operator return iterator not basic_iterator
a proper iterator's operator++(int) should return an instance of this
iterator, and  operator++() should return a reference of this
iterator. but chunked_fifo::iterator and chunked_fifo::const_iterator's
incremental operator returns basic_iterator<U>. this is not quite right
in the sense of the behavior of a proper iterator's interface.

so, in this change, instead of inheriting from the `basic_iterator`,
we just define the `iterator` and `const_iterator` as the instantiation
of this template.

Fixes #1793
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>