core: fix strerror_r usage from glibc extension
Curently seastar uses a glibc extension of strerror_r which does not exist in musl or POSIX.

The GNU-specific strerror_r() returns a pointer to a string containing the error message, which may or may not be in the buffer passed.

The posix XSI compiant version returns an int
https://linux.die.net/man/3/strerror_r

so check for the return value of strerror_r and if it returns non-zero, "unknown error" is returned, `buf` otherwise.