posix: use pthread_setaffinity_np instead of pthread_attr_setaffinity_np
pthread_attr_setaffinity_np is glibc specific and not portable. without which,
we cannot set the affinity when creating a pthread. but with another
extension pthread_setaffinity_np, we can set the CPU affinity after the thread
is created. hence this change. since both glibc and musl libc implements this
non-portable extension, we can use it here as an alternative of
pthread_attr_setaffinity_np.