tls: Move server name into tls_options
This commit moves the optional server name used for SNI into the
tls_options struct in order to make the API more consistent. All the
methods that create a TLS connection have gained a new overload which
accepts the tls_options struct as a param. The older methods have been
marked for deprecation.

Existing client code should still compile fine. If the caller specified
the name, the old overload is used. If the caller didn't specify the
name, the new overload is used.

I've also updated the tests and tls demo to use the non-deprecated
overload.