build: use stdout when defining enum only when possible
musl libc defines `stdout` like
```c++
  #define stdout (stdout)
```
while glibc defines `stdout` like
```c++
  #define stdout stdout
```
the same applies to `stderr`. both these definitions comply to the
the C++ standard, but we use `stdout` and `stderr` when defining
a `enum class`. so, after preprocessing, the definition from musl
libc is expanded into
```c++
enum class logger_ostream_type {
  none,
  (stdout),
  (stderr),
};
```
which does not compile. so, in order to be more standard compliant,
and does not rely on a specific libc implementation, while preserve
the backward compatibility. let's check if `stdout` and `stderr`
can be used as an identifier. also, since `stdout` and `stderr`
are required to be defined as macros by the standard library.
let's mark them deprecated in hope to remove them in a future
version in favor of `cout` and `cerr` which are guranteed to be
valid identifiers by the C++ standard.

but as the `[[deprecated]]` specifier cannnot be applied to a
certain enumerator, let's use the ` __attribute__ ((deprecated...))`
instead. fortunately, it is supported by both Clang and GCC.

this should address one of the FTBFS with musl libc.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>