util/process: s/stdout/cout/
musl libc defines `stdout` like
```c++
  #define stdout (stdout)
```
while glibc defines `stdout` like
```c++
  #define stdout stdout
```
the same applies to `stderr`. both these definitions comply to the
the C++ standard. if we use `stdout` or `stderr` in the place where
a C++ identifier is expected, we could run into the build failures
when compiling the source with musl libc, as we'd have the same build
failure we'd have when compiling:
```c++
class myclass {
    void foobaz() {
    }
};

int main() {
    myclass c;
    c.(foobaz)();
}
```
the compiling failure reads:
```
  <source>:9:7: error: expected unqualified-id before '(' token
    9 |     c.(foobaz)();
      |       ^
```
so, to address this problem, let's replace these names with less
dangerous ones. since `process` is still in `seasetar::experimental`
namespace, we can change it without in a faster pace.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>