http/api_docs: copy in api_docs's copy constructor
instead of hardwiring the apiVersion and swaggerVersion, we'd better
copy it from the the copied instance. also, it be more correct to
copy the "apis", which is the most important thing held by an
"api_docs".

this issue was spotted when auditing the code. as typically, we don't
use `api_docs` directly. instead, we use `api_registry_builder`, which
in general, registers the pointer to a newly created `api_registry` to
the specified `routes` instance. so we don't really copy an `api_docs`.

but this fix is still an improvement in the sense of correctness.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #1814