Prometheus: Disable Prometheus protobuf with a configuration
Prometheus Server's protobuf is back as an experimental option.

This patch adds a configuration option to Seastar Prometheus's
configuration object to control if Prometheus would answer protobuf
requests.

By default, it is off. It means that even if someone would set the
Prometheus server to use protobuf, without changing the configuration,
the result would still be in the text format.

To make it clearer, the bool text variable was renamed to is_text_format

Signed-off-by: Amnon Heiman <amnon@scylladb.com>