build: error out if find_program() fails
the programs being found are always required, so it would be better to error
out than moving on and ending with an obscure error message.
`find_programe()` does support the `REQUIRED` parameter, but it was introduced
by CMake 3.18, while our minimum required CMake version is 3.13. so let's check
it manually at this moment.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>