build: check if libucontext is necessary for using ucontext functions
musl libc does not offer the symbols of getcontext() and friends. it merely
provides the header file of `ucontext.h` with the declarations of these
functions so that "gcc with dwarf2 unwinding" can be built without errors.
these functions were removed from POSIX. fortunately, libucontext provides
a portable implementation.

in this change, if the function of `getcontext()` can not be compiled or
linked, we fall back to check libucontext, and use it as a fallback.

this should improve portability of Seastar, so it can compile on platforms
where glibc is not available.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>