scripts: addr2line: use yield instead defining a class
it is kind of overkill to define a class to implement an iterator.
less just use `yield` here. simpler this way, and less indent.

as we expect the resolver prints all decoded backtrace in a bulk,
we still materialize the input backtrace using `list()`, so the
input and output are not interleaved.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>