Do not add ext3 and ext4 to fs resolution map
We have a map of block magic to file system type and add ext2, 3 and 4
magic to this map. However, all of these have the same superblock so
the detected fs depends on the subtleties of braced map init when
there are duplicate keys. The standard is not clear but in practice
it seems like the first entry is used.

Avoid this by just including the ext2 mapping (the first one) and
adding a comment about why ext3 and ext4 are missing.