reactor: don't interact with the memory.cc when not using seastar the allocator
In particular:
* don't attempt to set reclaim hooks
* don't create poller for draining cross-cpu free list

Either of these will cause a SIGSEGV when seastar is configured to use
the standard memory allocator.

This patch is a follow-up to cbb0e888d81e8a2fa15493d9d85a0b84ca175914.
For some reason the crashes only reproduce when --smp > 1. That is why it
remained hidden so long

Fixes: scylladb/scylladb#12019