http/client: Mark connection as "keep in cache"
Currently if request finishes with unexpected status of handler callback
resolves with exception, the connection used to serve the request is not
returned back to pool and is instead closed.

This patch rewraps this logic by keeping the `bool _persistent` on the
connection itself. When putting the connection back this bit is checked
and the failure to handle the response set it. This is the same logic,
but allows for more ways to prevent connection from being cached in
pool in the future.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>