http/client: Add retry facility in case of connection failures
Currently if client::make_request() fails, the error is propagated back
to the caller whatever the error is. Sometimes the request fails because
of connection glitch, which can happen and in that case it's common to
try the same request again after a while in a hope that it was indeed
some teporary glitch.

The connection errors are told from others by checking the exception
being system_error with EPIPE or ECONABORTED codes.

This patch introduces an argument to client::make_request() that defines
how the restart should happen. There are two parameters in it

- the maximum number of retry attempts
- the base value for exponential-backoff-retry delays

By default the retries are off (max-count is zero), so callers will just
get EPIPE/ECONABORTED errors back.

The retry delay base is 100ms by default, so the caller that want some
sane retries may just specify the retry count and have it.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>