iostream: Configure batch flushes if sink can do it
Doing batch flushes carefully needs a certain support from the data sink
impl, in particular -- if batch flush fails in the background, the sink
should be able to handle it.

This patch adds the on_batch_flush() virtual method to the data sink
impl and enables batch flushes on output stream if requested AND sink
reports it can handle background failures.

This effectively turns off flush batching for all existing sinks, but it
will be fixed back by the next patch.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>