reactor: add cpu and awake time reactor metrics
Conceptually at all wall-clock moments the reactor is in one of three
states:

awake-busy
awake-idle
sleeping

Roughly speaking awake-busy is when the reactor has tasks or run or IO
to submit or reap, or there are waiting messages for this shard on the
cross-shard queues. This is the time counted as utilized in
"reactor utilization".

awake-idle means that the reactor is still running (or wanting to run
on the CPU), but does not have any work to do _now_, rather it spends
its time polling the cross-shard queues for new work, seeing if IO has
arrived to reap etc. From the OS point of view it is still demanding
CPU time just like the awake-busy state, but in the internal reactor
accounting we know it is not making forward progress on work, rather
just waiting for work to arrive.

sleeping - after some time in the awake-idle state and if certain
other conditions are met, the reactor may actually go to sleep. In this
state is uses an OS blocking call like io_getevents with a timeout to
deschedule itself from the CPU and be woken only when additional work
arrives (which will be detected somehow by the OS blocking mechanism).
In this state the reactor is not using additional CPU.

Having a clear view of these states in the reactor metrics is useful
and this change adds all three metrics with names similar to the above.

(cherry picked from commit 4c310e8ef90a9c3f7e4582cdd4f08539746842ce)