reactor: add better sleep time accounting
We track sleep time by taking timestamps before and after we start and
stop "sleep". This accounting is key for things like steal time, which
uses sleep time to subtract from real time to determine the time the
reactor wanted to run.

Currently we take the start timestamp far too early, and also in many
cases where we do not actually sleep: we take it before we "reap events
+ maybe sleep" but in many cases we do not sleep, yet count it as
sleep time. So we can do significant work in time which is being
recorded as "sleep".

On specific effect of this is that steal time can be negative, because
of the way steal time is calculated: inflated sleep time leads to
CPU time used (rusage) to be greater than reactor awake time (wall
clock based).

After this change, steal time can still be slightly negative but the
effect is reduced by 1 or 2 orders of magnitude.

(cherry picked from commit 0be77de0b36dc6ebee0ace15c696155733e1d6d0)