build: enable -Wdeprecated compiler option
so that for instance, we are aware that which pieces of code are

* using deprecated stuff with "deprecated" attribute attached to it
* using deprecated C++ feature

please note, this will enable a bunch of "-Wdeprecated-*" compiler
warning options, which are previously disabled. and these warning
will be taken as errors with the exception of "deprecated-declarations".
These warnings are from the `[[deprecated]]` attribute.

So this change requires us to address all of the other "-Wdeprecated-*"
warnings in Seastar. but this does not force the applications using
Seastar to do the same, as the "-Wdeprecated" option is a private CXX
flag which is only used when compiling Seastar itself.

Fixes #1863
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>