build: enable -Wdeprecated compiler option
so that for instance, we are aware that which pieces of code are

* using deprecated stuff with "deprecated" attribute attached to it
* using deprecated C++ feature

please note, this will enable a bunch of "-Wdeprecated-*" compiler
warning options, which are previously disabled. So this change
will enable the compiler to alert us whenever it identify deprecated
stuff when building Seastar. but this does not impact the applications
using Seastar, as the "-Wdeprecated" option is a private CXX flag
which is only used when compiling Seastar itself.

Fixes #1863
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>