tests: always wait for promise
in 24c6111c, we introduced a new test in rpc_test.cc, and it
creates a future, which is never waited. so, when the reactor
tears down, it clears it smp_message_queue, which could
contain the unresolved promise(s) associated with this future
if they are not resolved yet, but the way how
`promise_base::clear()` clears its state is to schedule its
pending task. so at that moment, the engine is not functional
anymore, and its `_task_queues` is empty. hence we have segfault
in the dtor of `engine` randomly when running
`test_rpc_metric_domains` test in `rpc_test`.

in this change, the promise is waited, as it is a part of the
test -- we take it into account when comparing the
"rpc_client_sent_messages" in "dom2".

please note, -Wunused compiler option is not able to identify
this problem.

Fixes #1868
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes scylladb/seastar#1869