memory: use `err` for the error string
in cf0c3ee0df1d64bb6e162d5b47edbba53c62215d, we detected for the GNU
variant of `strerror_r()` and falled back to the XSI variant if the
GNU version was not available. but we failed to handle the case
where `strerror_r()` succeeds in the XSI branch -- we should use the
buffer passed to `strerror_r()` for the returned `msg`. and `buf`
is not defined there. it was a copy-pasta error.

in this change, the error is corrected by s/err/buf/

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>