rpc: give compressors a way to send something over the connection
Fancy compressors might want to sends some information to the peer
(e.g. to offer a new dictionary) in some finite time.
But currently, there is no interface (or any mechanism) to make this
possible. (In particular, it's normal for a connection to be entirely
one-sided, giving the receiver no opportunity to communicate anything
to the sender).

This patch adds such a mechanism. A callable is passed to the compressors
on construction. Invoking this callable sends an empty frame to the peer.
Thus, if a compressor wants to send some metadata to the peer, it can
write down that this metadata should be prepended to the next message,
and then invoke the callable.