io_queue: do not dereference moved-away shared pointer
in dd6b20d484, we introduced flow monitor. in that commit, we reference
the `io_group` parameter after moving away from it. this breaks the
test when ASan is enabled. and more importantly, this introduced a
bug where a null pointer can be always dereferenced.

the seastar applications are now aborting when dereferecing a null pointer:

```
/home/kefu/dev/seastar/src/core/io_queue.cc:581:98: runtime error: member call on null pointer of type 'seastar::io_group'
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior /home/kefu/dev/seastar/src/core/io_queue.cc:581:98 in
Aborting on shard 0.
```

in this change, instead of dereferencing the moved-away pointer,
we deference the member variable instead. it is the destination of
the `std::move()`, so it should be valid.

Refs dd6b20d484
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes scylladb/seastar#1879