build: disable the -Wunused-* warnings for checkheaders
these warning does not help when checking the self-contain'ness
of these headers. actually they are always false alarms, as the way how
we check the self-contain'ness of headers is to rename the headers to
.cc files and compile them. but the symbols defined by the header
are not always used by the header itself. hence there are good chances
that the compiler would complain with these -Wunused-* warnings.

in this change, to silence these warnings,

- replace, for instance `-Wno-error=unused-variable` with
  `-Wno-unused-variable`, to completely disable these warnings.
- pass the options disabling these warning after the ones specified
  for the target being built. the options for building the target
  are likely to contain the option to enable these -Wunused-*
  warnings. if multiple options conflicts the last wins, so
  we need to put the -Wno-unused* options after them.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>