reactor: merge pollfn on I/O paths into a single one
this change consolidates the 3 pollfn classes into a single one.
this allows the backend to customize the way how the requests
are submitted, and how responses events are collected. for instance,
io_uring backend can submit and wait for events with a single syscall,
so a single place to do polling would make this possible.

in a follow-up change, we will move the part in the poller closer to
backend into backend.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>