condition-variable: make `waiter`s non-moveable
A lot of logic in waiters and waiter derived classes is creating data
structures or updating data structures living outside the `waiter`s that
refer to `this`. This is an easy way to shoot yourself in the foot if
the waiter is copied/moved after such reference is created (which I did
when trying to extend condition variable interface with `abort_source`
support -- my abort function would at some point try to abort a
destroyed waiter.)

It also allows writing code like this:
```
auto x = cv.when();

// Do something, potentially yield

co_await std::move(x);
```
which may or may not work, in any case it is definitely not being tested
in our tests. And supporting it requires more thought and care than
assuming we never yield before calling `co_await` on the waiter.

Improve robustness by explicitly deleting the move constructor (copy
constructor was already deleted).

It turns out that `predicate_awaiter` use cases depended on these
constructors. Adjust them so they don't.