Optimize zero size checking in allocator
When the allocator is handling a C++ allocation call (as opposed to
C's malloc), it must return a valid pointer, even for zero-size
allocation requests. The easy way to do this is to bump up any
sized requests to a larger value. We did this in various places
including in the fast path for the new-like functions.

Consolidate this and de-duplicate this check by using the existing
check where we bound below the size by sizeof(free_object) (8 bytes),
which saves 3-4 instructions on every path including the fast path.

There is one microscopic change in behavior: we now also set the size
non-zero for malloc calls to the underlying system allocator (which
happens, for example, for non-reactor threads), though glibc already
returns a non-null pointer for zero sized allocations so this change
doesn't seem material.